# static/docs/swagger.json
{"swagger":"2.0","info":{"title":"memory后端接口文档","version":"0.0.1","description":""},"paths":{"/register":{"POST":{"tags":["Users"],"summary":"用于注册用户 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"account","in":"path","required":true,"type":"string","description":"用户账户名 "},{"name":"password","in":"path","required":true,"type":"string","description":"密码 "},{"name":"mobile","in":"path","required":true,"type":"string","description":"手机号 "},{"name":"vip","in":"path","required":true,"type":"int","description":"是否注册Vip身份 0 普通用户 1 Vip用户 "},{"name":"recommend","in":"path","required":false,"type":"string","description":"邀请码 "}],"responses":{"200":{"description":"successful operation","schema":{"type":null,"items":{"$ref":"#/definitions/PostRegister"}}}}}},"/api/member/addMember":{"post":{"tags":["member"],"summary":"新增用户信息 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"新增用户信息 ","required":true,"schema":{"$ref":"#/definitions/addMember"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Number","items":{"$ref":"#/definitions/addMember"}}}}}},"/api/member/deleteMember":{"get":{"tags":["member"],"summary":"批量删除用户 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"array","description":"删除项目id[] "}],"responses":{"200":{"description":"successful operation","schema":{"type":"Number","items":{"$ref":"#/definitions/deleteMember"}}}}}},"/api/member/updateMember":{"post":{"tags":["member"],"summary":"更新用户信息 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"更新用户信息 ","required":true,"schema":{"$ref":"#/definitions/updateMember"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Number","items":{"$ref":"#/definitions/updateMember"}}}}}},"/api/user/doLogin":{"post":{"tags":["user"],"summary":"登录并获取用户信息 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"登录并获取用户信息 ","required":true,"schema":{"$ref":"#/definitions/doLogin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/doLogin"}}}}}},"/api/user/getUserinfo":{"get":{"tags":["user"],"summary":"获取用户信息 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"ACCESS-TOKEN-STAGE","in":"path","required":true,"type":"string","description":"用户cookies "}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/getUserinfo"}}}}}},"/api/user/setUserinfo":{"post":{"tags":["user"],"summary":"修改用户信息 ","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"修改用户信息 ","required":true,"schema":{"$ref":"#/definitions/setUserinfo"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Number","items":{"$ref":"#/definitions/setUserinfo"}}}}}}},"definitions":{"PostRegister":{"properties":{"account":{"type":"string","description":"用户账户名 "},"password":{"type":"string","description":"密码 "},"mobile":{"type":"string","description":"手机号 "},"vip":{"type":"int","description":"是否注册Vip身份 0 普通用户 1 Vip用户 "},"recommend":{"type":"string","description":"邀请码 "}},"required":["account","password","mobile","vip"]},"addMember":{"properties":{"email":{"type":"string","description":"邮箱 "},"name":{"type":"string","description":"姓名 "},"avatar":{"type":"string","description":"头像 "},"birthday":{"type":"string","description":"生日 "},"city":{"type":"string","description":"地址 "},"interest":{"type":"array","description":"爱好 "},"code":{"type":"number","description":"新增成功1 失败-1 "},"message":{"type":"string","description":"信息提示 "}},"required":["email","name","avatar","birthday","city","interest","code","message"]},"id":{"properties":{},"required":[]},"deleteMember":{"properties":{"code":{"type":"number","description":"删除成功1 失败-1 "},"message":{"type":"string","description":"信息提示 "}},"required":["code","message"]},"updateMember":{"properties":{"email":{"type":"string","description":"邮箱 "},"name":{"type":"string","description":"姓名 "},"avatar":{"type":"string","description":"头像 "},"birthday":{"type":"string","description":"生日 "},"city":{"type":"string","description":"地址 "},"interest":{"type":"array","description":"爱好 "},"code":{"type":"number","description":"更新成功1 失败-1 "},"message":{"type":"string","description":"信息提示 "}},"required":["email","name","avatar","birthday","city","interest","code","message"]},"doLogin":{"properties":{"email":{"type":"string","description":"用户邮箱 "},"password":{"type":"string","description":"用户密码 "},"code":{"type":"string","description":"是否登录成功1 失败-1 "},"message":{"type":"string","description":"信息提示 "}},"required":["email","password","code","message"]},"getUserinfo":{"properties":{"ACCESS-TOKEN-STAGE":{"type":"string","description":"用户cookies "},"email":{"type":"string","description":"邮箱 "},"name":{"type":"string","description":"姓名 "},"avatar":{"type":"string","description":"头像 "},"birthday":{"type":"string","description":"生日 "},"city":{"type":"string","description":"地址 "},"interest":{"type":"array","description":"爱好 "}},"required":["ACCESS-TOKEN-STAGE","email","name","avatar","birthday","city","interest"]},"setUserinfo":{"properties":{"email":{"type":"string","description":"邮箱 "},"name":{"type":"string","description":"姓名 "},"avatar":{"type":"string","description":"头像 "},"sex":{"type":"string","description":"性别 "},"birthday":{"type":"string","description":"生日 "},"city":{"type":"string","description":"所在地 "},"desc":{"type":"string","description":"描述 "},"interest":{"type":"array","description":"                                                    参数                           备注                                                                             interest                                                       { &quot;code&quot; : 1 ,&quot;message&quot; : &quot;修改成功&quot;}                             { &quot;code&quot; : 1 ,&quot;message&quot; : &quot;修改成功&quot;}                             { &quot;code&quot; : 1 ,&quot;message&quot; : &quot;修改成功&quot;}                             { &quot;code&quot; : 1 ,&quot;message&quot; : &quot;修改成功&quot;}                             { &quot;code&quot; : 1 ,&quot;message&quot; : &quot;修改成功&quot;}                             { &quot;code&quot; : 1 ,&quot;message&quot; : &quot;修改成功&quot;}                                                                            "},"code":{"type":"number","description":"修改成功1 失败-1 "},"message":{"type":"string","description":"信息提示 "}},"required":["email","name","avatar","sex","birthday","city","desc","interest","code","message"]}}}